+++ Cargo.toml
@@ -30,6 +30,7 @@ dirs = "4.0.0"
 futures-util = "0.3"
 indicatif = "0.16"
 reqwest = { version = "0.11", features = ["json", "blocking", "rustls-tls", "stream"], default-features = false }
+scraper = "0.15.0"
 serde = { version = "1.0", features = ["derive"] }
 serde_json = "1.0"
 serde_with = "2.0"
